update: makefile updated
